Update input portions of the spec (#62)
Updates most of the input portions of the spec with relevant algorithms from the comments in the explainer. It does not fully address currently existing issues (e.g. #57), and thus should mostly be a reflection of the current explainer.
Note that a few Gamepad specific algorithms still need to be expanded upon, but that may be done in the future.