Anchors extensions - allow the test to delay anchor creation
Current approach for anchor creation taken in the test API does not
allow the WPTs to delay anchor creation. Switching to a callback-based
approach with a callback returning a Promise<boolean> should allow us to
author tests that decide to hold off with anchor creation instead of
returning a result immediately.